Refactor RoundsAPage to integrate MarketDominanceCard and improve time formatting
- Replaced the previous DominanceChartSection with the new MarketDominanceCard component for enhanced market dominance display.
- Moved the formatRemainingTime function to a new utils file for better organization and reusability.
- Updated the layout in RoundsAPage for improved visual structure and user experience.
